refactor(components): 第二階層を責務ベースの 9 カテゴリへ再編
primitives / forms / layout の grab-bag を責務軸で切り直し、
identity / actions / inputs / navigation / feedback / overlays /
display / layout / utilities の 9 カテゴリに再編。

- 全 consumer は単一 barrel @dayopt/components から import するため
  公開 API・import 元は不変（barrel の相対 path のみ更新）
- cn.ts / hooks/ は src 直下据え置き（../cn 参照 33 箇所を不変に保つ）
- 移動で壊れる cross-import 6 行のみ修正
  (inputs/form, inputs/field, navigation/pagination,
   overlays/alert-dialog, actions/command)

typecheck / lint / lint:boundaries pass。
