+++ packages/components/src/utilities/visually-hidden.tsx
